fixed build errors with mongodb
